#VMS-491 System can now be bound to the cloud using desktop client
<==== Mercurial data ====>
Branch: nat_traversal

+++ nx_cloud/libcloud_db/src/stree/stree_manager.cpp
@@ -21,8 +21,8 @@ namespace nx {
 namespace cdb {
 
 QN_DEFINE_EXPLICIT_ENUM_LEXICAL_FUNCTIONS(StreeOperation,
-    ((int)StreeOperation::authentication, "authentication")
-    )
+    (StreeOperation::authentication, "authentication")
+)
 
 
 StreeManager::StreeManager() throw(std::runtime_error)
